Make sub-components and move scss to pattern
