+++ static/js/publisher/release/components/releasesConfirmDetails/types.js
@@ -1,0 +1,5 @@
+export default {
+  RELEASE: "release",
+  UPDATE: "update",
+  CANCELLATION: "cancel"
+};
