+++ static/sass/_snapcraft_release.scss
@@ -46,38 +46,6 @@
     padding-bottom: .5rem;
   }
 
-  .p-releases-confirm-details__row {
-    align-items: center;
-    display: grid;
-    font-size: .875rem;
-    grid-column-gap: 1rem;
-    grid-template-columns: 4rem 15rem 2rem auto 16rem;
-    padding: 1rem;
-
-    &.is-closing {
-      grid-template-columns: auto;
-    }
-
-    &:nth-child(odd) {
-      background: $color-light;
-    }
-
-    &-join {
-      text-align: center;
-    }
-
-    &-progress {
-      display: grid;
-      font-weight: 400;
-      grid-column-gap: .5rem;
-      grid-template-columns: auto 8rem;
-    }
-
-    &-notes {
-      color: $color-dark;
-    }
-  }
-
   // RELEASES TABLE
 
   .p-releases-table {
