+++ static/sass/styles.scss
@@ -227,3 +227,6 @@ $color-social-icon-foreground: $color-light;
 
 @import 'snapcraft_p-progressive-bar';
 @include snapcraft-p-progressive-bar;
+
+@import 'snapcraft_p-release-details-row';
+@include snapcraft-p-release-details-row;
